simplified code
merged some methods necessary for file handling in new class FileManagement, in package mgmt.
