--- src/main/java/util/DictionaryGenerator.java
@@ -1,46 +1,0 @@
-package util;
-
-import java.io.BufferedReader;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.HashSet;
-
-public class DictionaryGenerator{
-	public static HashSet<String> readFile(String fileName) {
-		Path file = Paths.get(fileName);
-		HashSet<String> dictionary = new HashSet<>();
-		try {
-			BufferedReader reader = Files.newBufferedReader(file, StandardCharsets.UTF_8);
-			String sCurrentLine = null;
-			while ((sCurrentLine =reader.readLine()) != null) {
-				dictionary.add(sCurrentLine);
-			}
-			reader.close();
-		} catch (FileNotFoundException e) {
-			e.printStackTrace();
-		} catch (IOException e) {
-			e.getMessage();
-		}
-		return dictionary;
-	}
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
